requirements.txt
fastapi==0.109.0
passlib==1.7.4
pydantic==2.5.3
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0.1
SQLAlchemy==2.0.25
starlette==0.35.1
uvicorn==0.25.0
python-multipart==0.0.6
